Started in on TwoDS.  Other cleanup.
Moved some ProcessRecord method static vars to call members.

Started using isSyncWord() and implemented isOverloadWord()

Introduced clockMhz member to Probe for probes that need a timeWord divisor.

First pass at TwoDS::ProcessRecord(), still needs work.

Work in MainCanvas to start plotting TwoDS.

Improved/added comments.

+++ xpms2d/class/Particle.h
@@ -22,20 +22,20 @@ public:
   long time;           // Seconds since mid-night
   long msec;
 
-  size_t w, h;
-  size_t area;
-  bool  vert_gap;               // Was there a blank slice (i.e. multiple parts)
-  bool  horiz_gap;              //   and in the hozizontal dir?
+  size_t w, h;		// Particle width and height
+  size_t area;		// Total diodes shadowed by this particle
+  bool  vert_gap;	// Was there a blank slice (i.e. multiple parts)
+  bool  horiz_gap;	//   and in the hozizontal dir?
   bool  reject;
-  unsigned char edge;           // particle touched either edge
-  bool  timeReject;             // Nimbus will reject (because of bad timing bar)
+  unsigned char edge;	// particle touched either edge
+  bool  timeReject;	// Nimbus will reject (because of bad timing bar)
 
-  ushort        x1, x2;         // for particles that touch both edges.
+  ushort        x1, x2;	// for particles that touch both edges.
 
 // Is this timeWord size sufficient for the Fast2D???
-  uint32_t      timeWord;
-  uint32_t      deltaTime;      // Amount of time between prev & this particle
-  uint32_t      liveTime;       // Amount of time consumed by particle
+  uint64_t      timeWord;
+  uint32_t      deltaTime;	// Amount of time between prev & this particle
+  uint32_t      liveTime;	// Amount of time consumed by particle
 };
